+++ moxy-compiler/src/main/java/moxy/compiler/viewstate/ViewInterfaceProcessor.kt
@@ -2,6 +2,7 @@ package moxy.compiler.viewstate
 
 import com.squareup.javapoet.JavaFile
 import moxy.compiler.ElementProcessor
+import moxy.compiler.MvpCompiler
 import moxy.compiler.MvpCompiler.Companion.elementUtils
 import moxy.compiler.MvpCompiler.Companion.messager
 import moxy.compiler.asDeclaredType
@@ -13,11 +14,13 @@ import moxy.viewstate.strategy.AddToEndSingleStrategy
 import moxy.viewstate.strategy.StateStrategyType
 import java.util.*
 import javax.lang.model.element.AnnotationMirror
+import javax.lang.model.element.Element
 import javax.lang.model.element.ElementKind
 import javax.lang.model.element.ExecutableElement
 import javax.lang.model.element.Modifier
 import javax.lang.model.element.TypeElement
 import javax.lang.model.type.TypeKind
+import javax.lang.model.type.TypeMirror
 import javax.tools.Diagnostic.Kind
 
 class ViewInterfaceProcessor(
@@ -27,23 +30,22 @@ class ViewInterfaceProcessor(
 ) : ElementProcessor<TypeElement, ViewInterfaceInfo> {
 
     private val frameworkDefaultStrategy: TypeElement
-    private var viewInterfaceElement: TypeElement? = null
-    private var viewInterfaceName: String? = null
+    private lateinit var viewInterfaceElement: TypeElement
+    private lateinit var viewInterfaceName: String
     private val migrationMethods = mutableListOf<MigrationMethod>()
 
     init {
         if (defaultStrategy != null) {
-            var localDefaultStrategy: TypeElement? =
-                elementUtils.getTypeElement(defaultStrategy)
+            var localDefaultStrategy: TypeElement? = elementUtils.getTypeElement(defaultStrategy)
             if (localDefaultStrategy == null) {
-                val message =
-                    String.format("Unable to parse option %s. Check %s exists", defaultStrategy, defaultStrategy)
-                messager.printMessage(Kind.ERROR, message)
-                localDefaultStrategy = NEW_DEFAULT_STATE_STRATEGY
+                messager.printMessage(
+                    Kind.ERROR,
+                    "Unable to parse option '$OPTION_DEFAULT_STRATEGY'. Check $defaultStrategy exists")
+                localDefaultStrategy = DEFAULT_STATE_STRATEGY
             }
             frameworkDefaultStrategy = localDefaultStrategy
         } else {
-            frameworkDefaultStrategy = NEW_DEFAULT_STATE_STRATEGY
+            frameworkDefaultStrategy = DEFAULT_STATE_STRATEGY
         }
     }
 
@@ -51,29 +53,34 @@ class ViewInterfaceProcessor(
         this.viewInterfaceElement = element
         this.viewInterfaceName = element.simpleName.toString()
 
-        val methods = mutableListOf<ViewMethod>()
-
-        val interfaceStateStrategyType = getInterfaceStateStrategyType(element)
-
-        // Get methods for input class
-        getMethods(element, interfaceStateStrategyType, ArrayList(), methods)
-
-        // Add methods from super interfaces
-        methods.addAll(iterateInterfaces(0, element, interfaceStateStrategyType, methods, ArrayList()))
+        // Get methods from input interface
+        val methods = getMethods(element)
 
         // Allow methods to have equal names
         val methodsCounter = mutableMapOf<String, Int>()
         for (method in methods) {
-            var counter = methodsCounter[method.name]
-            if (counter != null && counter > 0) {
+            val counter = methodsCounter[method.name] ?: 0
+            if (counter > 0) {
                 method.uniqueSuffix = counter.toString()
-            } else {
-                counter = 0
             }
-            counter++
-            methodsCounter[method.name] = counter
+            methodsCounter[method.name] = counter + 1
         }
-        return ViewInterfaceInfo(element, methods)
+
+        return ViewInterfaceInfo(element, methods.toList())
+    }
+
+    /**
+     * Returns ViewMethod for each suitable method from this interface and its superinterfaces
+     */
+    private fun getMethods(element: TypeElement): Set<ViewMethod> {
+        // Get methods from input class
+        val enclosedMethods = getEnclosedMethods(element)
+
+        // Get methods from super interfaces
+        val methodsFromSuperinterfaces = iterateInterfaces(element)
+
+        // Combine and exclude overridden methods
+        return combineMethods(enclosedMethods, methodsFromSuperinterfaces)
     }
 
     fun makeMigrationHelper(): JavaFile? {
@@ -84,42 +91,28 @@ class ViewInterfaceProcessor(
         }
     }
 
-    private fun getMethods(
-        typeElement: TypeElement,
-        defaultStrategy: TypeElement?,
-        rootMethods: List<ViewMethod>,
-        superinterfacesMethods: MutableList<ViewMethod>
-    ) {
-        for (element in typeElement.enclosedElements) {
-            // ignore all but non-static methods
-            if (element.kind != ElementKind.METHOD || element.modifiers.contains(Modifier.STATIC)) {
-                continue
+    /**
+     * Returns ViewMethod for each suitable enclosed method into this interface (but not superinterface)
+     */
+    private fun getEnclosedMethods(viewInterface: TypeElement): Set<ViewMethod> {
+        val defaultStrategy = getInterfaceStateStrategyType(viewInterface)
+        return viewInterface.enclosedElements
+            .filter {
+                // ignore all but non-static methods
+                it.kind == ElementKind.METHOD && !it.isStatic()
             }
-
-            val methodElement = element as ExecutableElement
-
-            val method = getViewMethod(methodElement, typeElement, defaultStrategy)
-
-            if (rootMethods.contains(method)) {
-                continue
-            }
-
-            if (superinterfacesMethods.contains(method)) {
-                checkStrategyAndTagEquals(method, superinterfacesMethods[superinterfacesMethods.indexOf(method)])
-                continue
-            }
-            superinterfacesMethods.add(method)
-        }
+            .map { getViewMethod(it as ExecutableElement, viewInterface, defaultStrategy) }
+            .toSet()
     }
 
     private fun getViewMethod(
         methodElement: ExecutableElement,
-        typeElement: TypeElement,
+        viewInterface: TypeElement,
         defaultStrategy: TypeElement?
     ): ViewMethod {
         if (methodElement.returnType.kind != TypeKind.VOID) {
-            val message = "You are trying to generate ViewState for ${typeElement.simpleName}. " +
-                    "But ${typeElement.simpleName} contains non-void method \"${methodElement.simpleName}\" " +
+            val message = "You are trying to generate ViewState for ${viewInterface.simpleName}. " +
+                    "But ${viewInterface.simpleName} contains non-void method \"${methodElement.simpleName}\" " +
                     "with the return type of ${methodElement.returnType}."
             messager.printMessage(Kind.ERROR, message, methodElement)
         }
@@ -134,7 +127,7 @@ class ViewInterfaceProcessor(
         } else {
             if (defaultStrategy == null && !disableEmptyStrategyCheck) {
                 if (enableEmptyStrategyHelper) {
-                    migrationMethods.add(MigrationMethod(typeElement, methodElement))
+                    migrationMethods.add(MigrationMethod(viewInterface, methodElement))
                 } else {
                     val message = ("A View method has no strategy! " +
                             "Add @StateStrategyType annotation to this method, or to the View interface. " +
@@ -145,67 +138,60 @@ class ViewInterfaceProcessor(
             defaultStrategy ?: frameworkDefaultStrategy
         }
 
-        // get tag from annotation
         val tagFromAnnotation = annotation?.getValueAsString(StateStrategyType::tag)
         val methodTag: String = tagFromAnnotation ?: methodElement.simpleName.toString()
 
         return ViewMethod(
-            viewInterfaceElement!!.asDeclaredType(),
+            viewInterfaceElement.asDeclaredType(),
             methodElement,
             strategyClass,
             methodTag)
     }
 
-    private fun checkStrategyAndTagEquals(
-        method: ViewMethod,
-        existingMethod: ViewMethod
-    ) {
-        val differentParts: MutableList<String> = ArrayList()
-        if (existingMethod.strategy != method.strategy) {
-            differentParts.add("strategies")
-        }
-        if (existingMethod.tag != method.tag) {
-            differentParts.add("tags")
-        }
-        if (differentParts.isNotEmpty()) {
-            val arguments = method.parameters.joinToString(", ")
-            val parts = differentParts.joinToString(" and ")
-            throw IllegalStateException(
-                "Both ${existingMethod.enclosedClassName} and ${method.enclosedClassName} " +
-                        "has method ${method.name}($arguments) with different $parts. " +
-                        "Override this method in $viewInterfaceName or make $parts equal"
-            )
-        }
-    }
-
+    /**
+     * Returns ViewMethods for all superinterfaces, or empty set if element does not have superinterfaces
+     */
     private fun iterateInterfaces(
-        level: Int,
-        parentElement: TypeElement,
-        parentDefaultStrategy: TypeElement?,
-        rootMethods: List<ViewMethod>,
-        superinterfacesMethods: MutableList<ViewMethod>
-    ): List<ViewMethod> {
-        for (typeMirror in parentElement.interfaces) {
-            val anInterface = typeMirror.asTypeElement()
-
-            val typeArguments = typeMirror.typeArguments
-            val typeParameters = anInterface.typeParameters
-
-            require(typeArguments.size <= typeParameters.size) {
-                "Code generation for the interface ${anInterface.simpleName} failed. Simplify your generics."
+        viewInterface: TypeElement
+    ): Set<ViewMethod> {
+        return viewInterface.interfaces
+            .map {
+                getTypeAndValidateGenerics(it)
             }
-
-            val defaultStrategy = parentDefaultStrategy ?: getInterfaceStateStrategyType(anInterface)
-
-            getMethods(anInterface, defaultStrategy, rootMethods, superinterfacesMethods)
-
-            iterateInterfaces(level + 1, anInterface, defaultStrategy, rootMethods, superinterfacesMethods)
-        }
-        return superinterfacesMethods
+            .flatMapToSet {
+                // implicit recursion
+                getMethods(it)
+            }
     }
 
-    private fun getInterfaceStateStrategyType(typeElement: TypeElement): TypeElement? {
-        val annotation = typeElement.getAnnotationMirror(StateStrategyType::class)
+    private fun getTypeAndValidateGenerics(interfaceMirror: TypeMirror): TypeElement {
+        val superinterface = interfaceMirror.asTypeElement()
+
+        val typeArguments = interfaceMirror.typeArguments
+        val typeParameters = superinterface.typeParameters
+
+        require(typeArguments.size <= typeParameters.size) {
+            "Code generation for the interface ${superinterface.simpleName} failed. Simplify your generics."
+        }
+
+        return superinterface
+    }
+
+    /**
+     * Combines methods, comparing by [ViewMethod.equals], discarding duplicates from superinterface
+     */
+    private fun combineMethods(
+        methods: Set<ViewMethod>,
+        superInterfaceMethods: Set<ViewMethod>
+    ): Set<ViewMethod> {
+        return methods + superInterfaceMethods
+    }
+
+    /**
+     * Returns default StateStrategyType for this [viewInterface], if specified
+     */
+    private fun getInterfaceStateStrategyType(viewInterface: TypeElement): TypeElement? {
+        val annotation = viewInterface.getAnnotationMirror(StateStrategyType::class)
         val value = annotation?.getValueAsTypeMirror(StateStrategyType::value)
         return if (value != null && value.kind == TypeKind.DECLARED) {
             value.asTypeElement()
@@ -214,8 +200,17 @@ class ViewInterfaceProcessor(
         }
     }
 
+    private fun Element.isStatic(): Boolean {
+        return modifiers.contains(Modifier.STATIC)
+    }
+
+    private inline fun <T, R> List<T>.flatMapToSet(transform: (T) -> Iterable<R>): Set<R> {
+        return flatMapTo(LinkedHashSet(), transform)
+    }
+
     companion object {
-        private val NEW_DEFAULT_STATE_STRATEGY: TypeElement =
+        private const val OPTION_DEFAULT_STRATEGY = MvpCompiler.DEFAULT_MOXY_STRATEGY
+        private val DEFAULT_STATE_STRATEGY: TypeElement =
             elementUtils.getTypeElement(AddToEndSingleStrategy::class.java.canonicalName)
     }
 }